fix : 150줄 넘는 컴포넌트만 styles.tsx 파일로 분리 처리
