+++ src/pages/index.tsx
@@ -1,8 +1,9 @@
 import React from 'react'
 import ReactFullpage, { fullpageApi } from '@fullpage/react-fullpage'
-import MainSection from '../components/main'
-import Header from '../components/header'
-import { DownBtn } from './styles'
+import MainSection from '../components/Main'
+import Header from '../components/Header'
+import styled from '@emotion/styled'
+import { mq } from '../utils/media-query'
 
 type Props = {
   fullpageApi: fullpageApi
@@ -73,3 +74,29 @@ const Fullpage: React.FC = () => {
 }
 
 export default Fullpage
+
+const DownBtn = styled.button`
+  transform: translateX(-50%) !important;
+  left: 50%;
+  position: absolute;
+  bottom: 5.2083vw;
+  border: none;
+  background: none;
+  border-radius: 50%;
+  border: 1px solid #fff;
+  cursor: pointer;
+
+  ${mq({
+    width: ['60px', '40px;', '45px;', '45px;', '60px;', '60px;'],
+    height: ['60px', '40px;', '45px;', '45px;', '60px;', '60px;'],
+  })};
+  img {
+    width: 70%;
+    transform: rotate(-90deg);
+  }
+  &.top {
+    img {
+      transform: rotate(90deg);
+    }
+  }
+`
